Use floats for comparison to avoid issues with converting to int

+++ pkg/streamingpromql/optimize/plan/remove_statically_empty_expressions.go
@@ -141,7 +141,7 @@ func isAlwaysEmpty(node planning.Node, params *planning.QueryParameters) bool {
 }
 
 func isAlwaysEmptyBinaryExpression(node *core.BinaryExpression, params *planning.QueryParameters) bool {
-	earliestPossibleTimestampValueInMilliseconds := params.TimeRange.StartT - params.LookbackDelta.Milliseconds()
+	earliestPossibleTimestampValueInMilliseconds := float64(params.TimeRange.StartT - params.LookbackDelta.Milliseconds())
 
 	if node.ReturnBool {
 		return false
@@ -155,25 +155,25 @@ func isAlwaysEmptyBinaryExpression(node *core.BinaryExpression, params *planning
 		// timestamp(v) < C: always empty when C <= the earliest value that timestamp() could return
 		// timestamp() returns the value in seconds since the epoch, so we need to convert to milliseconds.
 		if constant, ok := isTimestampComparison(node.LHS, node.RHS); ok {
-			return int64(constant*1000) <= earliestPossibleTimestampValueInMilliseconds
+			return constant*1000 <= earliestPossibleTimestampValueInMilliseconds
 		}
 
 	case core.BINARY_LTE:
 		// timestamp(v) <= C: always empty when C < the earliest value that timestamp() could return
 		if constant, ok := isTimestampComparison(node.LHS, node.RHS); ok {
-			return int64(constant*1000) < earliestPossibleTimestampValueInMilliseconds
+			return constant*1000 < earliestPossibleTimestampValueInMilliseconds
 		}
 
 	case core.BINARY_GTR:
 		// C > timestamp(v): equivalent to timestamp(v) < C.
 		if constant, ok := isTimestampComparison(node.RHS, node.LHS); ok {
-			return int64(constant*1000) <= earliestPossibleTimestampValueInMilliseconds
+			return constant*1000 <= earliestPossibleTimestampValueInMilliseconds
 		}
 
 	case core.BINARY_GTE:
 		// C >= timestamp(v): equivalent to timestamp(v) <= C.
 		if constant, ok := isTimestampComparison(node.RHS, node.LHS); ok {
-			return int64(constant*1000) < earliestPossibleTimestampValueInMilliseconds
+			return constant*1000 < earliestPossibleTimestampValueInMilliseconds
 		}
 	}
